[8.2.0] Include `DefaultInfo` in `cquery` `providers` dict (#25719)
All targets observed from Starlark advertise the `DefaultInfo` provider,
but internally that provider is split into `FilesProvider`,
`FilesToRunProvider` and `RunfilesProvider`. This is confusing to users
that use `cquery` to inspect targets using Starlark expressions. This is
improved by synthesizing a `DefaultInfo` provider in the same way as
during regular analysis. As a result, the `default_runfiles` and
`data_runfiles` of a target can now be inspected with `cquery`.

Closes #25702.

PiperOrigin-RevId: 741291403
Change-Id: Ieeb92afb47cd3215fe8dde15c305831806c7bf87

Commit
https://github.com/bazelbuild/bazel/commit/c3661a2471e79cfc189d9dc6bd383084d4e13e6e

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>