Don't hard code StalarkSemantics into ParamDescriptor
Previously, the state of `semantics` (if the parameter was enabled or disabled) was indirectly hardcoded into the ParamDescriptor.

Accurately encode the conditions described in the annotation in ParamDescriptor.
Check at evaluation time with current thread's semantics, if the parameter is enabled.

After this change ParamDescriptor doesn't depend on StarlarkSemantics. This is better for caching.

Issue: https://github.com/bazelbuild/bazel/issues/25743
PiperOrigin-RevId: 742265869
Change-Id: I09f2184389b34ffcbc394b7e1c469c4059dc1007