Allow git_repository to reference directories relative to the declaring module
git_repository's implementation is such that repositories are fetched by creating a
new repository in a working directory, adding the value of the remote kwarg as a git remote
and fetching that remote.

This works well for repositories which are git/http/ssh urls, and absolute local paths.

However, using a relative directory (including ., the current directory) is precluded by the
change directory step in the repository implementation.

This commit allows users to use git_repository with a relative directory, unblocking novel
usage patterns like the one in the integration test.

The integration test example fetches a different branch of the current repository as a remote.

This could be used to ensure cross-release compatibility, allowing tests to depend on resources
from v1 and v2 branches of a repository and ensure that they are equivalent or compatible in some way.