Invalidate action cache entries when archived artifacts are toggled on or off.
Otherwise, the action cache entry might be missing the archived representation when one is expected (making toggling the flag ineffectual) or, conversely, contain one when not expected (increasing memory consumption, although that's a minor concern). Currently, this causes a precondition check in ActionOutputMetadataStore.injectTree() to fail, which is what motivated this fix.

Although it would be technically possible to salvage the entry and still get a cache hit in some cases, it's not worth the complexity as we don't expect this flag to be frequently toggled.

Also remove the precondition check from ActionOutputMetadataStore.injectTree(), because even with archived tree artifacts enabled, it's still possible for the archived representation to be missing if the action ran locally (currently, that never results in metadata injection since the action cache only stores remote metadata, but that could change in the future).

PiperOrigin-RevId: 743063947
Change-Id: Ifb77e151cab466bb6e25ba871a47b020eaeb6fc2