Plumb the metadata for files referenced by Filesets through ExpandedArtifact.
The benefit is that these won't be considered local files anymore for the purposes of uploading files. In addition to a minor performance improvement, this also allows us to remove the mutation of ActionInputMap from ArtifactValueFileSystem.

That, in turn, will make it possible for ArtifactValueFileSystem to use an InputMetadataProvider instead of an ActionInputMap, which in turn enable putting a Skyframe-based InputMetadataProvider in ActionFileSystem, which will then unlock removing the logic to do Skyframe restarts from ActionFileSystem.

RELNOTES: None.
PiperOrigin-RevId: 743648043
Change-Id: Ie61fbbb1f107b106f56d36ab90786411b83e45dc