[8.2.0] Enable sparse checkout in git_repository rule (#25765)
Add a sparse_checkout_patterns and sparse_checkout_file arguments to the
git_repository rule. If non-empty, this will cause the rule to perform a
sparse checkout of the repository, only checking out files that match
the .gitignore-style patterns in this list.

Fixes issue https://github.com/bazelbuild/bazel/issues/24069.

Cherry-pick of
https://github.com/bazelbuild/bazel/commit/4006e059737d214fc4116083d9c4e58d5a2e40c0