Implement a general interface for function types
Refactor Type evaluation to type resolution and move into resolver.

Create a new package for types. They are not pure syntax, but they are needed by the resolver.

Function types play an important role in type checking and it's also important to have a nice interface that's complete and easy to handle.

This is an internal interface, that will in the future progress into `Callable` type, but at the moment it's not exposed. It will be used for runtime/dynamic type checking.

The design causes a small regression in memory. I'd prefer not to do premature optimizations here, because they might make the implementation that follow much more complex.

Issue: https://github.com/bazelbuild/bazel/issues/25520
PiperOrigin-RevId: 743895944
Change-Id: I0ec661a12c842912e60b174bb9f64cc0cbad1320