Remove support for legacy struct providers
The internal cleanup has finished and the flag is flipped in Bazel 8.

Keep the same error messages and behavior. To make this change smaller I temporarily introduced DefaultInfo.createEmpty. It serves to trigger the same code path as before and to produce error with the same line numbers.

Update tests. Remove tests that already have non-struct version. Rewrite others with non-struct version.

PiperOrigin-RevId: 743933503
Change-Id: I9ea90fa9c6735dc85fa3824367b638d02c04972d