Fix a memory leak in `BuildIntegrationTestCase` by calling `BlazeModule#blazeShutdown` after every test.
In `BuildIntegrationTestCase`, `BlazeModule` instances are thrown away in favor of new ones for every test case. This does not happen on a real blaze server, where there is only a single instance of each `BlazeModule` for the server's lifetime. Thus, `BlazeModule` implementations may leak resources when they are dereferenced without explicit cleanup.

PiperOrigin-RevId: 743979176
Change-Id: I7510147c4a575827cf49ef7706f270cde5d83cee