[8.2.0] Set generator_name, generator_function, generator_location, full Starlark stack for rules in symbolic macros (#25772)
Before, we did not set generator_* pseudo-attributes for rules generated
in a symbolic macro. However, it turns out that there are existing tools
which expect to be able to filter rule targets by generator_name in
query expressions; and such tools get broken when legacy macros are
migrated to symbolic. The easiest solution is to set generator_name and
friends consistently for both legacy and symbolic macro rule targets.

In the long term, we wish to deprecate and remove generator_*
pseudo-attrs, since the API has numerous problems and doesn't expose the
information users really need:

But for now, however, we ought to support the broken API symmetrically
for all varieties of macro.

In addition, take the opportunity to concatenate the Starlark stacks of
BUILD and symbolic macro threads for rule targets in symbolic macros,
because (1) the generator_* psudo-attributes are derived from the BUILD
thread's stack, and (2) the stack is shown in some query output, and
truncating it at symbolic macro boundaries is likely to confuse users
who rely on it for debugging.

RELNOTES: Set generator_name, generator_function, generator_location,
and the full Starlark stack for rule targets instantiated in a symbolic
macro.

PiperOrigin-RevId: 744110163
Change-Id: I5216b05f22ee0663343bd4468f4f299d095bb739

Commit
https://github.com/bazelbuild/bazel/commit/d6b07eb19209a5426594ca642dfe657d98c624f4