Use BAZEL_LLVM_PROFDATA instead of GCOV to specify the coverage tool for LLVM coverage
Counterintuitively, the GCOV variable is used in collect_cc_coverage.sh
instead of the BAZEL_LLVM_PROFDATA variable to specify the llvm-profdata tool
when performing LLVM specific coverage processing.

This is a distinct case from using LLVM tools to perform "gcov compatible"
coverage, in which case continuing to use "GCOV" is appropriate.

Closes #25516.

PiperOrigin-RevId: 745139853
Change-Id: I82c7e324817bf92ce2e4770c7eb26920ef00a634