Stop using a dirty bit in PersistentMap.
Instead, save can simply attempt to flush the in-memory journal, which is more obviously safe against concurrent updates. The one exception is clear(), which bypasses the journal and requires a full save to work correctly (we previously got this wrong and did an incremental save).

Also rename a few PersistentMap methods to clarify their purpose.

PiperOrigin-RevId: 745593372
Change-Id: Ic3967b74a19c13f874385bd6d27ec5cda86a6d76