Have `RunfilesTree#getMapping` return a `SortedMap`
It already always returns a `SortedMap` in production code, but signaling this on the type level makes it safer to consume the map in contexts where sorting is required, e.g. when computing Merkle trees.

Also drop an impossible switch case in adjacent code that is flagged by an IDE lint.

Closes #25625.

PiperOrigin-RevId: 747396796
Change-Id: I3be6e1320c4325fc0c52119cdda56b0073bb7934