Make test toolchain requirement mandatory
This ensures that a test for which no compatible exec platform is available results in an analysis error (improved by the test toolchain type's `no_match_error`) instead of being run on the first execution platform, which is likely to fail with an uninformative "exec format" error.

RELNOTES[inc]: The toolchain requirement of the default `test` exec group is now mandatory.