[8.3.0] macOS: respect DEVELOPER_DIR for Xcode selection. (#25885)
On macOS, this change ensures that when `DEVELOPER_DIR` environment
variable is used, it will be propagated to `/usr/bin/xcrun` to query the
SDKROOT. This takes precedence over the `--xcode_version` flag when
selecting the Xcode installation. This aligns with the intended use as
documented in `man xcrun`:

`The tool xcode-select(1) is used to set a system default for the active
developer directory, and may be overridden by the DEVELOPER_DIR
environment variable (see ENVIRONMENT).`

To leverage this, developers must propagate the `DEVELOPER_DIR`
environment variable to Bazel actions using
`--action_env=DEVELOPER_DIR`.

Closes #25657.

PiperOrigin-RevId: 748404233
Change-Id: I52181098658d33e2bd2ec992ad1cd38d64abb293

Commit
https://github.com/bazelbuild/bazel/commit/d2e8c82570c63e9a0c6ef369ef458a86f587e819

Co-authored-by: Benji Vos <bvos@apple.com>