[8.3.0] Fix crash on lost inputs that are symlinks (#25856) (#25890)
The downloads in `AbstractActionInputPrefetcher#prefetchInputs` reported
lost inputs that are symlinks to other artifacts with the exec path of
the target, which results in a crash when the target is not also an
input to the action. This is fixed by always reporting the exec path of
the symlink.

Fixes #25841

Closes #25847.

PiperOrigin-RevId: 747819887
Change-Id: Iff9125c95a2598851b8a31ed4ca18fd7218a97aa (cherry picked from
commit a7da0d8e693b6c73a99a8bb7ebff6611fbd994ab)

Closes #25849

Commit
https://github.com/bazelbuild/bazel/commit/74da5dd52ecb2a7a3573121ea028992e8e827723

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>