Build execlog_parser.jar in compile.sh
Execlog parser can be useful to provide in distributions along
with bazel binary itself.

At the same time, Bazel isn't guaranteed to be buildable with itself,
only `.bazelversion` is guaranteed to build given revision, or the
bootstrapped bazel.

So it seems that the right place to build execlog_parser is within
compile.sh (or require using .bazelversion which is less convenient
requirement for packaging use-case)