Allow PackageFunction to compute PackagePiece.ForBuildFile
PackageFunction, in addition to producing a PackageValue (containing a Package)
given a PackageIdentifier, will now produce a PackagePieceValue.ForBuildFile
(containing a PackagePiece.ForBuildFile) if given a
PackagePieceIdentifier.ForBuildFile.

We reuse PackageFunction instead of adding a new skyfunction because the logic
for producing a PackageValue and a PackagePieceValue.ForBuildFile is virtually
identical; furthermore, PackageFunction already has its own class hierarchy with
three implementations, and adding a further level of hierarchy for full vs.
piecewise evaluation seems like a recipe for unreadability.

Introduce PackageoidValue as a superclass of PackageValue and PackagePieceValue
for convenience of skyframe's label interning logic and tests.

Refactor PackageFunctionTest to test PackageFunction's full-package and lazy
macro expansion modes.

We'll need a separate skyfunction to produce package pieces for macros - will be
added in a followup.

Working towards https://github.com/bazelbuild/bazel/issues/23852

PiperOrigin-RevId: 750720626
Change-Id: I7b33ad0e8f8ae2bfa795459a047da96498176f98