Add profiler span for sending the `Execute` request
We are seeing profiles in which there is a large gap between the start of the `execute remotely` span and the the start of the `queue` span reported by the remote executor. The current theory is that there is contention on the channel with async tasks spawned by `--remote_cache_async` which this span could help pin down.