Simplify convert_library_to_link_list_to_linker_input_list
Works towards removing LegacyLinkerInputs.
The cl improves readability and my understanding what the function does.

The change is no-op.

It might be hard to see why this is true for moving ` || !supportsDynamicLinker` from inner to outer `if`. Before, I believe that in inner `if` when this part of the condition is true, libraryInputToUse is overwritten with static library if static exists. After, in the outer `if` when this part of conditions is true, libraryInputToUse is set to static library, and not overwritten with an interface library.

PiperOrigin-RevId: 751451082
Change-Id: Iafd8399dcd45f6b7aa9cd2cc63cab7ea6388928d