Use LTO mapping instead of converting LegacyLinkerInputs
_lto_indexing_action used to construct new LegacyLinkerInput with object files replaced with bitcode files.
lto_mapping is a preexisting map, that is used to replace bitcode files with object files in the final linking step.
Instead of constructing new objects, we use lto_mapping to do the substitution.

Works towards removing LegacyLinkerInputs.

The change doesn't set must_keep_debug to False like previous code. Because of the complexity of the code there might be other differences.

PiperOrigin-RevId: 751455084
Change-Id: Ibc3c41aa578990acf5140c7a75c0c0db9aa9351f