Make ActionFileSystem take an InputMetadataProvider instead of an ActionInputMap.
The end goal is to remove the part of ActionFileSystem that reaches into Skyframe and replace that with the same logic in SkyframeInputMetadataProvider. However, to do so, it looks like some logic still needs to be fixed and we need to make sure that action rewinding still works when generated inputs are lost at an inopportune time.

RELNOTES: None.
PiperOrigin-RevId: 751729860
Change-Id: I8e89188246ac317be0f054e38f57183a3f89620e