Add rewinding test coverage for a test action losing an input, with assertions on BEP events.
Surprisingly, we have no test coverage for a test action observing a lost input. The test is added to `TargetSummaryEventTest` (not `RewindingTest`) for two reasons:

1. `TargetSummaryEventTest` goes through `BlazeCommandDispatcher`, which actually executes `TestCommand`. This is necessary for exercising `AggregatingTestListener`, which powers BEP-relevant test events.
2. `TargetSummaryEventTest` is a good place to assert on `TargetSummary` and `TestSummary` BEP events.

Helper methods for extracting BEP events are made to use `collect(toOptional())`, which asserts that there is no more than one element in the stream. The `TargetSummary` method helpfully dumps the full BEP if there is no such event (which was very helpful for debugging a bazel-only test failure).

PiperOrigin-RevId: 752701852
Change-Id: I5a8a068ca0f0084c84883fb661368e9035e8f7ac