Make ActionFileSystem use SkyframeInputMetadataProvider and stop ActionFileSystem in input discovery.
These two steps need to be done together because before this change, ActionFileSystem relied on all input metadata lookups going through is to that learns what input files can conceivably be used, which in turn allowed it to not throw a MissingDepExecException after input discovery.

This makes SkyframeInputMetadataProvider do the fallback Skyframe input metadata lookup instead of ActionFileSystem itself. This allows one to remove that functionality from ActionFileSystem, which will happen in a follow-up change.

There are a number of other cleanups that this change unlocks (for example, creating the ActionInputMap/SkyframeInputMetadatProvider input metadata provider just once), but let's keep this change small; it's a pretty delicate one.

RELNOTES: None.
PiperOrigin-RevId: 752726012
Change-Id: I6528d3b2db396cd2fcdbed1b8dd7f1fde17e7d7a