Implement an ActionCache trim() method.
This is the method responsible for garbage collecting entries that haven't been used in a while. Garbage collection operates by copying all entries below a given cutoff age into a new cache; atomically replacing the old on-disk representation with the new; then reloading the cache from the new representation. This requires minimal additional logic, is safe in the event of interruption (either intentionally due to a subsequent command, or unintentionally due to a crash), and performs well enough even for the largest caches in use at Google.

The code is currently unused; a followup CL will wire it into an idle task.

Progress on #22515.

PiperOrigin-RevId: 753153798
Change-Id: Ia2d7e2dbe96586635ed0c9255e6538ecb96a3783