Propagate root causes when error bubbling propagates an error from a dependency pruned by the starlark `unused_inputs_list` feature.
unknown commit propagated the exception but did not propagate the root causes. This confuses `CompletionFunction`, which assumes that an `ActionExecutionException` has non-empty root causes. When root causes are empty and `!env.valuesMissing()`, `CompletionFunction` posts a successful `TargetCompleteEvent` reporting the broken top-level artifacts as built. In the case of a Fileset, this causes a crash during BEP reporting with `--experimental_build_event_expand_filesets` enabled because the Fileset's metadata is missing.

PiperOrigin-RevId: 753394189
Change-Id: I70d6c9f9ec9c3202dffef862da4a0fde9f032953