Attempt invalidation lookup on all existing deserialized nodes before analysis.
This is a key step towards incremental builds with a remote analysis cache service.

This introduces a new method in `RemoteAnalysisCachingDependenciesProvider` to invalidate `SkyValue`s using their fingerprinted `SkyKey`s. Doing the remote lookup may add more latency on every build before analysis. Current experiments show that the performance may not be a huge concern (sub-second invalidation for O(10000) keys).

This also requires a few changes to the `SkyframeExecutor`, because it needs to persist the deserialized keys across command invocations for invalidation.

PiperOrigin-RevId: 753993489
Change-Id: I06cb244c04e469ba3bcfd226348a7523601e8d6a