Replace `NestedSet#memo` with a `WeakReference` to the cached list representation.
The primary motivation for this change is to reduced retained heap by switching from the strongly-referenced `byte[] memo` to a `WeakReference<ImmutableList<E>>`. However, benchmarks show that this is also a win for cpu/wall, plus getting rid of the bit-packed memo and the tricky synchronization around it is a nice complexity reduction.

Experiments show that this change increases the number of "full" NestedSet traversals, but it's a net win because followup traversals are now usually "free" (read a `WeakReference`) instead of merely "cheaper" (pruned traversal using `memo`). The `WeakReference` approach is effective because repeated `NestedSet#toList` calls on a particular instance are typically close to each other temporally (during a particular action's execution), so there is unlikely to be intervening GC. Additionally, the "full" traversal is slightly less costly because we don't need to write `memo`. Savings may also come from reduced memory pressure.

A previous improvement was made in b/314282963 to clear `NestedSet#memo` for an action's inputs after the action executes. While this approach is effective, it does not scale to all `NestedSet` use cases. For example, the current memory inefficiency  motivating this change occurs during Starlark command line expansion, where the `NestedSet` instance is not an action's inputs. I elected to keep the eager clearing in order to drop the `WeakReference`, as keeping it around did not reduce the number of "full" traversals.

PiperOrigin-RevId: 755382268
Change-Id: I889980b2f708e47cafb08745a1a9cb957b1b1e42