Apply visibility break-glass flags in a more consistent manner
The prerequisite validator has the ability to turn off visibility checking, either globally or just for the experimental package tree. Previously, this logic was run after visibility checking was complete, when deciding whether or not to display an error in the case of a visibility violation.

This CL moves these checks up sooner. This lets us simplify the job of `handleVisibilityConflict()` (renamed `reportVisibilityConflict()`).

It also lets us apply the experimental package tree loophole to the location that's actually being checked against the dependency target's visibility, rather than the location of the consuming target. For example, if a symbolic macro in the non-experimental package tree has a visibility violation, but the macro was called from within the experimental tree, we would previously not show an error when visibility checking for experimental was disabled. Added a test case for this.

Also moved the package_group prerequisite check into its own top-level call from `validate()`, since it's not really part of the visibility check proper.

Work toward #25933.

PiperOrigin-RevId: 755464892
Change-Id: Ia43ace65d64139af9dc2d54b023ec3455b634db8