Implement an idle task to garbage collect the action cache.
The `--experimental_action_cache_gc_max_age` flag determines the criteria for considering an action cache entry stale; zero disables garbage collection. The default will be replaced with a suitable nonzero value in a future CL.

In addition, two flags are provided to specify the idle time (`--experimental_action_cache_gc_idle_delay`) and the percentage of entries that must be stale (`--experimental_action_cache_gc_threshold`) in order for a garbage collection to occur. Their default values are meant to avoid repeatedly recreating the action cache inside a rebuild loop or when there's little to be collected.

Progress on #22515.

PiperOrigin-RevId: 755829702
Change-Id: I2ecbd3d585baf72b30da0ba398718130e5a9e4d6