Make the default GC idle task its own class and register it explicitly.
Even though the default task is documented as running after other tasks, this is not enforced and appears to never have been. Since other tasks have a fairly large default delay (on the order of minutes) the documented behavior no longer seems reasonable.

Treating the default idle task identically to other tasks makes it easier to collect stats for all of them (in a followup).

PiperOrigin-RevId: 756285912
Change-Id: I034320c400843990af84aeb9f15bdabd41006abe