Memoize Skyframe lookups in SkyframeInputMetadataProvider.
With action rewinding, Skyframe can "lose" SkyValues at any given time. The code is carefully laid out such that when action rewinding can happen, the *first* metadata request for a given input only ever happens through .getInputMetadataChecked() and this worked because ActionFileSystem memoized the lookups.

SkyframeInputMetadataProvider did not do so, thus if an action was rewound after a given direct reverse dependency did the first lookup for one of its outputs but before a subsequent one for the same output, it ended up throwing a MissingDepExecException, which in turn got wrapped into an IllegalStateException by .getInputMetadata().

RELNOTES: None.
PiperOrigin-RevId: 757819478
Change-Id: Icb7d34052011a3a860b5f9524d8e8c7b3be412b7