Add VisibilityCheckState to prepare for adding more detailed error diagnostics
This simplifies information sharing between visibility helper functions. In a future CL, it should make it easier for the checking logic to propagate delegation information back to the error reporting logic.

In the current CL, I use this to make the code for handling implicit dep visibility more readable.

Also added TODOs for this visibility work and removed some unneeded BUILD deps.

In validateDirectPrerequisiteVisibility, renamed a lengthy boolean name to "isImplicitDep" and flipped its usage in the if-else block. Renamed reportVisibilityConflict to generateVisibilityConflictMessage, and eliminated its direct usage of the rule context.

Work toward #25933.
Work toward #25941.

PiperOrigin-RevId: 757840605
Change-Id: I80accbd1feb8167f6449a2c7101a439c6e31ef84