Skycache: Fix missed cache hits due to mutated frozen MutableStarlarkList
When we serialize in a writer build, the mutability field has already been mutated and has the value IMMUTABLE, then when we shutdown or do clean expunge followed by a reader build, when we serialize the mutability field hasn't been mutated yet because the mutation happens during analysis (in createConfiguredTarget). What we are trying to do is to serialize to get the key and see if we can skip analysis altogether so at that point we haven't run createConfiguredTarget yet.

If after that we do another reader build but we don't shutdown nor clean --expunge (which also shuts down), then MutableStarlarkList.mutability will remain as IMMUTABLE from the previous invocation (after analysis was run). Since this matches the values serialized at the end of the writer build, we will get more cache hits than a reader build with a cold server.

This is because the specific MutableStarlarkList that was causing the problem came from a build option and a reference to it stayed in the BuildOptions cache. The BuildOptions object is part of a SkyKey. SkyKeys and its constituents are supposed to be immutable throughout the lifetime of the object, but we don't have tests for that.

RELNOTES:none
PiperOrigin-RevId: 758193992
Change-Id: I6eb30733827ea139a249b60c2db45f7ce127bc18