Replace `--define=EXECUTOR=remote` with toolchains and selects
This allows for cross-platform builds to work seamlessly without a need to manually toggle between prebuilt and from source toolchains.

As a positive side effect, this change also paves the way for dropping a dependency on an exec-configured C++ toolchain of Java and Python targets built for non-Windows platforms. Rulesets will need to migrate to using the launcher maker toolchain instead of the `launcher_maker` target. rules_shell wasn't affected by this as it already defined its own launcher toolchain.

Work towards #19587

Closes #26019.

PiperOrigin-RevId: 758697189
Change-Id: Ibfa272390056963ef35301614520fd38d018210e