Refactor isVisibleToDeclaration to use RuleOrMacroInstance
`isVisibleToDeclaration`'s argument type is changed from `Object` to the more accurate `RuleOrMacroInstance`. This requires lifting `getPackageMetadata` and `getDeclaringMacro` up to `RuleOrMacroInstance`.

Also renamed the "declaration" terminology of this method's name and argument name to "consumer", since I felt "declaration" to be redundant / not descriptive in this context.

The various target and macro related classes are by now a sea of confusing method names related to packages and containing macros. Added a TODO to clean that up at some indeterminate point in the future.

Work toward #25933.

PiperOrigin-RevId: 758804676
Change-Id: Id585598f2a20c26e9126ef44c1e09be4fadd2bc8