Serialize `ResolvedSymlinkArtifactValue#resolvedPath` as a relative path by finding its containing root.
In practice, this absolute path points to a source artifact - it's used when a `SymlinkAction` creates a symlink to a source artifact. Storing the resolved path as a `PathFragment` as opposed to an `Artifact` makes it easier to encode it in the local action cache. We could store a relative path fragment if builds with multiple source roots go away.

PiperOrigin-RevId: 758900056
Change-Id: I5e4e04717fa717d0b0e107c8b7a666e74335fac4