Fix `CcCompilationHelper.shouldProvideHeaderModules()` and its uses.
The problem is that `shouldProvideHeaderModules()` evaluates whether `publicHeaders` and `privateHeaders` are empty, and it is called in `addHeader()` and `addPrivateHeader()` which add elements to `publicHeaders` and `privateHeaders`, respectively, before respectively after calling `shouldProvideHeaderModules()`.

The check for whether any headers are present makes sense in `createCcCompileActions()` where it is called to check whether a header module would be empty (and thus useless) due to no headers, and thus shouldn't be created.

In `addHeader()` and `addPrivateHeader()` the call to `shouldProvideHeaderModules()` serves just the purpose to avoid creating header actions to syntax check the header because that would be covered by a header module anyway; this decision is independent of the number of headers already added to the compilation.

Therefore, in `addHeader()` and `addPrivateHeader()` the call to `shouldProvideHeaderModules()` is replaced with a new `isHeaderModulesEnabled()` call which doesn't look at the number of already added headers.

The test `testHeaderParsingUsingHeaderModules()` whether a parse headers action would be based on a header module in case header modules are enabled; this behaviour is no longer desired, and the test only passed because for private headers, even with header modules enabled, a parse header action was created for the *first* private header added to a compilation, because in `addPrivateHeader()` the call to `shouldProvideHeaderModules()` is done *before* adding the header, so for the first header `privateHeaders.isEmpty()` is true, and a parse header action is created.

PiperOrigin-RevId: 759038165
Change-Id: Ia3b30370544b2106d83e6974bc64ba01fa8fe311