Automated rollback of commit 665aa04fad18e0830d8cff79616b4916010668af.
*** Reason for rollback ***

Breakages in CI

*** Original change description ***

Serialize `ResolvedSymlinkArtifactValue#resolvedPath` as a relative path by finding its containing root.

In practice, this absolute path points to a source artifact - it's used when a `SymlinkAction` creates a symlink to a source artifact. Storing the resolved path as a `PathFragment` as opposed to an `Artifact` makes it easier to encode it in the local action cache. We could store a relative path fragment if builds with multiple source roots go away.

PiperOrigin-RevId: 759040775
Change-Id: I7e8d4c62922069249f229e5d00036c83a9b3db73