Simplify RemoteActionFileSystem.downloadIfRemote().
Notice that the call to prefetchFiles is only provided with inputArtifactData, and it skips files for which metadata isn't available; therefore nothing can break if downloadIfRemote ignores inputs from other sources.

It would arguably be more correct to make getInputStream() work for remote outputs as well, as there's no written rule that it must only be called on inputs; however, I'll defer that change for now since everything works without it.

PiperOrigin-RevId: 759189365
Change-Id: I9412aa35cbec660dbc8a906e14b076f1a4d40a4b