Record the reason for "no match" in config matching providers
This prepares for allowing consuming rules, selects as well as toolchain resolution to report why a particular condition didn't match.

I'm submitting this separately as it unlocks a number of other improvements to error messages, but should first be vetted for its potential impact on memory usage.

Work towards #11984
Work towards #22850

Closes #26035.

PiperOrigin-RevId: 759190944
Change-Id: I9246cec662c0d07e838e29df5fd5819b0ce7d5a6