Mark SkyframeLookups abandoned if there are errors or state eviction.
SkyframeLookups are ListenableFutures and deserialization may add
listeners to them. However, the future requires a corresponding
Skyframe lookup to be performed to complete. The SkyframeLookups are
sometimes retained in SkyKeyComputeState and can be evicted. They are
sometimes skipped because there is a previous error. This change marks
SkyframeLookups abandoned in both of those cases, allowing listeners
to progress, rather than hanging.

In the case where the SkyframeLookupCollector fails fast, marks any
enqueued lookups abandoned, and marks any subsequently added lookups
abandoned without enqueuing.

PiperOrigin-RevId: 759614205
Change-Id: I6e03bcbcd9dd289211979efa9901aab539ca191a