Split out collect_solib_dirs, update and optimize Starlark code
The remaining native code LibrariesToLinkCollector deals with collecting solib directories. This cl enables Starlark code doing the same, that was previously written. Optimize it a little, so that there's no regression.

This way the worse tech-debt in the linking code is isolated into a single file. Once time permits, it should be rewritten from scratch.

PiperOrigin-RevId: 759646479
Change-Id: I581d14cde3da369fe3942f6e76454e3349c0738a