Close `--system` files left open by `javac` in `JavaBuilder`
When using `--system` with a modularized JDK, two of its files are kept open by the JDK after the JavacTask has completed:

* `lib/jrt-fs.jar` is loaded by the host JDK via a `URLClassLoader` that is never closed explicitly.
* `lib/modules` is memory-mapped by the `JrtFileSystem` implementation provided by the system JDK, but never explicitly unmapped (see https://bugs.openjdk.org/browse/JDK-4724038).

As long as these files are kept open, Bazel can't delete them on Windows, which causes failures when using a sandboxed execution mode (e.g. a sandboxed multiplexed worker) with the reduced classpath optimization, which may delete the input files immediately after compilation when it needs to fall back to the full classpath.

As a workaround, these resources are closed explicitly in `JavaBuilder`.

A reference to the `URLClassLoader` can be obtained from any `Path` backed by the file system. We use the location of the `java.base` module in the `--system` JDK to obtain the `FileSystemProvider` and thus its class loader.

Obtaining and closing the mapped `modules` `ByteBuffer` requires intrusive reflection as well as a call to `Unsafe#invokeCleaner`. All non-`Error` `Throwable`s encountered along the way are ignored to guard against breakages with other JDKs.