Announce combined instead of individual baseline coverage in the BES
This simplifies the bookkeeping for baseline coverage files and avoids the
creation of actions with `O(n^2)` total inputs.

Work towards #5716

RELNOTES[INC]: Baseline coverage files for individual tests are no longer announced in the BES. Instead, a new `baseline_report.lcov` file containing the merged baseline coverage information for the entire invocation is now announced in the BES. It is also still merged into the combined report.

Closes #25274.

PiperOrigin-RevId: 761078796
Change-Id: I28bf8efaf5b09610034b5042a7516339a2fce7ad