Skycache: fix incorrect invalidation when syncing from one primed evaluating version to another primed evaluating version.
The `AnalysisCacheInvalidator` wasn't accounting for the version of the keys for the invalidation lookup, causing a failure mode where keys of deserialized nodes from an old version is still getting cache hits on the newer version, when they were supposed to be invalidated.

Make it such that if the lookup is for a different version, all of the existing deserialized nodes should be invalidated immediately. There is no need to consult the AnalysisCacheService then.

PiperOrigin-RevId: 761458908
Change-Id: Ia73996f24f2620f4f960ff0c1e3943697572f9cf