Add diagnostic message for visibility errors that could've been fixed with delegation
This CL adds a message (gated on --verbose_visibility_errors) identifying a macro caller who could've seen the dependency, and suggests that maybe the caller (or a child) could've passed the dependency in as an argument.

One weakness is that we identify the caller using isVisibleToLocation(), which itself does not consider delegation. So if A -> B -> C, where A passes the dependency to B but B does not pass it to C, we report that A can see it and suggest that A or something underneath it (really, B) pass the dep onward. Ideally, we would recognize that B can see it due to delegation, and suggest specifically that B pass it on to C.

Also:
- Remove spurious newlines from some bullets.
- Add TODO about how we can't easily report children that could've been exported to their caller, unless we add macro info to the TargetData abstraction.
- Add a test that we do not get a spurious same-package bullet where it trivially doesn't apply.

Work toward #25933.

PiperOrigin-RevId: 761537385
Change-Id: I9138e339eb4ca4dc9f160d813921a9b46f09ea41