Add message clarifying behavior of aliases in visibility errors
With --verbose_visibility_errors enabled, this gives an additional bullet point reminding the user that it doesn't matter what the visibility of a target is when you refer to that target via an alias.

There are a couple of TODOs remaining for bullet points I didn't add because it would require extending (or working around) the `TargetData` abstraction. These are things that would require reading the actual attributes or macro ancestry information of the dependency target. I'm not prioritizing that work now, so closing the tracking bug.

Fixes #25933.

PiperOrigin-RevId: 761549968
Change-Id: Id8d04db02f3338ee7ee462799fe8362c76fafd13