Allow marking ParameterFileWriteAction as executable.
This is used by ctx.actions.write(...) when its given an "args" object
to write out. We should still respect the "is_executable" flag in this
case.

The observable change is that the output file has the executable bit set
and the mnemonic is "Writing script" instead of "Writing file".

PiperOrigin-RevId: 761894384
Change-Id: I731e84b93b0d6bf1b4c584b2f3f41319663aab2c