Download full tree artifact if matched by `--remote_download_regex`
This makes it easier to download a full tree artifact and avoids the confusing situation in which one is matched but the resulting directory is empty. It also allows for more efficient regular expressions.

Also adds an interesting test case that broke while working on https://github.com/bazelbuild/bazel/pull/25976, which originally contained this change.

Closes #26135.

PiperOrigin-RevId: 762421175
Change-Id: Ibb74c43748315602b60ddbb5280a5d617c3416ca