Avoid digesting files in `AbstractActionInputPrefetcher` if possible
Remote metadata that is later materialized as a local file frequently hits the code path in `AbstractActionInputPrefetcher` that checks for discrepancies between the metadata and the local filesystem state. Since computing the digest is potentially costly and often unnecessary in this case, short-circuit by comparing sizes and contents proxies, if available.

Also add comments to `Path#statNullable` and `Path#statIfFound` to more clearly explain the subtle difference.

Closes #26146.

PiperOrigin-RevId: 763396727
Change-Id: I081d6d8b4efba661348746ba56c39120d18c31c3