skycache: fix download builds when an alias depends on a `RemoteConfiguredTargetValue`.
Producing a `RemoteConfiguredTargetValue` will intentionally not trigger an evaluation of its `Package` dep, because it's not necessary. However, this breaks `ConfiguredTargetAndData` instantiation from existing Skyframe nodes, because the `PackageValue` will be null.

Fix that by adding an interface method on `ConfiguredTargetValue` to return its `TargetData` projection directly, which is what `ConfiguredTargetAndData` needs. This is only non-null for remote CTs.

We could also do an `instanceof RemoteConfiguredTargetValue` check, but that adds a circular dependency because `RCTV` depends on `ConfiguredTargetAndData` already.

PiperOrigin-RevId: 763425171
Change-Id: I900710bd38e63c85f9c5d40e3994858b4a671fcb