Set `TARGET_CPU` make variable with the name of the cpu constraint value of the target platform.
This change is guarded by `--incompatible_target_cpu_from_platform`. `--experimental_override_name_platform_in_output_dir` is renamed to `--experimental_override_platform_cpu_name` and now also used to provide a mapping from the platform label to the value of `TARGET_CPU` for that platform. If no mapping is provided for the current platform, the name of the cpu constraint value label is used.

PiperOrigin-RevId: 763495306
Change-Id: Ic7f88b8228849cd568f221e7db64c6a9c5abcea9